Cast to int()

+++ lib/js/Resources.py
@@ -324,7 +324,7 @@ class Resources:
         info = self.getInfo()
         code = json.dumps(info, separators=(',',':'))
         
-        logging.info("Generated %sKB of resource info" % (len(code)/1024))
+        logging.info("Generated %sKB of resource info" % (int(len(code)/1024)))
         
         return "(function(){this.%s=%s})();" % (to, code)
